Expand test_customer_views and the testing framework. Remove MutableList.
Tests are now able to use a real Postgres database. In the tests,
transactions never commit and the schema is re-created for each
test module. Tests should use a real database because we need
to thoroughly test how the code uses the database. (External
integration tests will never test enough SQL/SQLAlchemy edge cases.)

Converted a test to the use the real database and added tests of the
'send' API call because it needed to change with the removal of
the MutableList column type.

The MutableList column type was interesting, but it's less risky in
SQLAlchemy to treat column values as immutable. Moved the mutation
logic to the 'send' API call.